fix: resolve CI failures for Hytte-aabh
Move parseTargetTime out of StridePage.tsx into strideUtils.ts to satisfy
react-refresh/only-export-components (files may only export React components).
