+++ web/src/pages/strideUtils.ts
@@ -1,0 +1,16 @@
+// Parses a strict "H:MM:SS" target-time string to seconds. Returns null for
+// empty/whitespace-only input or any string that does not match the format
+// exactly (non-negative hours, minutes 0–59, seconds 0–59). Two-part inputs
+// like "25:00" are intentionally rejected to avoid the historical ambiguity
+// where they were silently treated as hours+minutes and saved as 25 hours.
+export function parseTargetTime(s: string): number | null {
+  const trimmed = s.trim()
+  if (!trimmed) return null
+  const match = trimmed.match(/^(\d+):(\d{1,2}):(\d{1,2})$/)
+  if (!match) return null
+  const hours = Number(match[1])
+  const minutes = Number(match[2])
+  const seconds = Number(match[3])
+  if (minutes > 59 || seconds > 59) return null
+  return hours * 3600 + minutes * 60 + seconds
+}
